+++ client/doc_test.go
@@ -3,22 +3,50 @@ package client
 import (
 	"context"
 	"log"
-	"os"
 	"time"
 
+	"github.com/gammazero/nexus/router"
 	"github.com/gammazero/nexus/wamp"
 )
 
+func ExampleConnectLocal() {
+	// Create router that local client attaches to.
+	routerConfig := &router.Config{
+		RealmConfigs: []*router.RealmConfig{
+			{
+				URI:           wamp.URI("nexus.realm1"),
+				AnonymousAuth: true,
+			},
+		},
+	}
+	r, err := router.NewRouter(routerConfig, nil)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	// Configure and connect local client.
+	c, err := ConnectLocal(r, Config{Realm: "nexus.realm1"})
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer c.Close()
+}
+
+func ExampleConnectNet() {
+	// Configure and connect client.
+	c, err := ConnectNet("unix:///tmp/app.sock", Config{Realm: "nexus.realm1"})
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer c.Close()
+}
+
 func ExampleClient_Call() {
 	// Configure and connect caller client.
-	logger := log.New(os.Stdout, "", 0)
-	cfg := Config{
-		Realm:  "nexus.realm1",
-		Logger: logger,
-	}
+	cfg := Config{Realm: "nexus.realm1"}
 	caller, err := ConnectNet("ws://localhost:8080/", cfg)
 	if err != nil {
-		logger.Fatal(err)
+		log.Fatal(err)
 	}
 	defer caller.Close()
 
@@ -30,24 +58,20 @@ func ExampleClient_Call() {
 	callArgs := wamp.List{1, 2, 3, 4, 5, 6, 7, 8, 9, 10}
 	result, err := caller.Call(ctx, "sum", nil, callArgs, nil, "")
 	if err != nil {
-		logger.Fatal(err)
+		log.Fatal(err)
 	}
 
 	// Print the result.
 	sum, _ := wamp.AsInt64(result.Arguments[0])
-	logger.Println("The sum is:", sum)
+	log.Println("The sum is:", sum)
 }
 
 func ExampleClient_CallProgress() {
 	// Configure and connect caller client.
-	logger := log.New(os.Stdout, "", 0)
-	cfg := Config{
-		Realm:  "nexus.realm1",
-		Logger: logger,
-	}
+	cfg := Config{Realm: "nexus.realm1"}
 	caller, err := ConnectNet("ws://localhost:8080/", cfg)
 	if err != nil {
-		logger.Fatal(err)
+		log.Fatal(err)
 	}
 	defer caller.Close()
 
@@ -55,7 +79,7 @@ func ExampleClient_CallProgress() {
 	progHandler := func(result *wamp.Result) {
 		// Received more progress ingo from callee.
 		percentDone, _ := wamp.AsInt64(result.Arguments[0])
-		logger.Printf("Test is %d%% done", percentDone)
+		log.Printf("Test is %d%% done", percentDone)
 	}
 
 	// Create a context to cancel the call in one minute if not finished.
@@ -67,29 +91,25 @@ func ExampleClient_CallProgress() {
 	result, err := caller.CallProgress(
 		ctx, "run.test", nil, wamp.List{time.Second * 2}, nil, "", progHandler)
 	if err != nil {
-		logger.Fatal("Failed to call procedure:", err)
+		log.Fatal("Failed to call procedure:", err)
 	}
 
 	// As a final result, in this example, the callee returns the boolean
 	// result of the test.
 	passFail, _ := result.Arguments[0].(bool)
 	if passFail {
-		logger.Println("Test passed")
+		log.Println("Test passed")
 	} else {
-		logger.Println("Test failed")
+		log.Println("Test failed")
 	}
 }
 
 func ExampleClient_Register() {
 	// Configure and connect callee client.
-	logger := log.New(os.Stdout, "", 0)
-	cfg := Config{
-		Realm:  "nexus.realm1",
-		Logger: logger,
-	}
+	cfg := Config{Realm: "nexus.realm1"}
 	callee, err := ConnectNet("tcp://localhost:8080/", cfg)
 	if err != nil {
-		logger.Fatal(err)
+		log.Fatal(err)
 	}
 	defer callee.Close()
 
@@ -105,7 +125,7 @@ func ExampleClient_Register() {
 
 	// Register procedure "sum"
 	if err = callee.Register("sum", sum, nil); err != nil {
-		logger.Fatal("Failed to register procedure:", err)
+		log.Fatal("Failed to register procedure:", err)
 	}
 
 	// Keep handling remote procedure calls until router exits.
@@ -114,14 +134,10 @@ func ExampleClient_Register() {
 
 func ExampleClient_Register_progressive() {
 	// Configure and connect callee client.
-	logger := log.New(os.Stdout, "", 0)
-	cfg := Config{
-		Realm:  "nexus.realm1",
-		Logger: logger,
-	}
+	cfg := Config{Realm: "nexus.realm1"}
 	callee, err := ConnectNet("tcp://localhost:8080/", cfg)
 	if err != nil {
-		logger.Fatal(err)
+		log.Fatal(err)
 	}
 	defer callee.Close()
 
@@ -147,7 +163,7 @@ func ExampleClient_Register_progressive() {
 
 	// Register example procedure.
 	if err = callee.Register("system_test", sendData, nil); err != nil {
-		logger.Fatal("Failed to register procedure:", err)
+		log.Fatal("Failed to register procedure:", err)
 	}
 
 	// Keep handling remote procedure calls until router exits.
@@ -156,14 +172,10 @@ func ExampleClient_Register_progressive() {
 
 func ExampleClient_SendProgress() {
 	// Configure and connect callee client.
-	logger := log.New(os.Stdout, "", 0)
-	cfg := Config{
-		Realm:  "nexus.realm1",
-		Logger: logger,
-	}
+	cfg := Config{Realm: "nexus.realm1"}
 	callee, err := ConnectNet("ws://localhost:8080/", cfg)
 	if err != nil {
-		logger.Fatal(err)
+		log.Fatal(err)
 	}
 	defer callee.Close()
 
@@ -189,7 +201,7 @@ func ExampleClient_SendProgress() {
 
 	// Register example procedure.
 	if err = callee.Register("system_test", sendData, nil); err != nil {
-		logger.Fatal("Failed to register procedure:", err)
+		log.Fatal("Failed to register procedure:", err)
 	}
 
 	// Keep handling remote procedure calls until router exits.
@@ -204,27 +216,23 @@ func ExampleClient_Subscribe() {
 	)
 
 	// Configure and connect subscriber client.
-	logger := log.New(os.Stdout, "", 0)
-	cfg := Config{
-		Realm:  "nexus.realm1",
-		Logger: logger,
-	}
+	cfg := Config{Realm: "nexus.realm1"}
 	subscriber, err := ConnectNet("ws://localhost:8080/", cfg)
 	if err != nil {
-		logger.Fatal(err)
+		log.Fatal(err)
 	}
 	defer subscriber.Close()
 
 	// Define specific event handler.
 	handler := func(args wamp.List, kwargs wamp.Dict, details wamp.Dict) {
 		eventData, _ := wamp.AsString(args[0])
-		logger.Printf("Event data for topic %s: %s", site1DbErrors, eventData)
+		log.Printf("Event data for topic %s: %s", site1DbErrors, eventData)
 	}
 
 	// Subscribe to event.
 	err = subscriber.Subscribe(site1DbErrors, handler, nil)
 	if err != nil {
-		logger.Fatal("subscribe error:", err)
+		log.Fatal("subscribe error:", err)
 	}
 
 	// Define pattern-based event handler.
@@ -233,21 +241,21 @@ func ExampleClient_Subscribe() {
 		// subscriptions contain the matched topic in the details.
 		eventTopic, _ := wamp.AsURI(details["topic"])
 		eventData, _ := wamp.AsString(args[0])
-		logger.Printf("Event data for topic %s: %s", eventTopic, eventData)
+		log.Printf("Event data for topic %s: %s", eventTopic, eventData)
 	}
 
 	// Subscribe to event with prefix match.
 	options := wamp.Dict{"match": "prefix"}
 	err = subscriber.Subscribe(site2DbAny, patHandler, options)
 	if err != nil {
-		logger.Fatal("subscribe error:", err)
+		log.Fatal("subscribe error:", err)
 	}
 
 	// Subscribe to event with wildcard match.
 	options["match"] = "wildcard"
 	err = subscriber.Subscribe(site1AnyAlerts, patHandler, options)
 	if err != nil {
-		logger.Fatal("subscribe error:", err)
+		log.Fatal("subscribe error:", err)
 	}
 
 	// Keep handling events until router exits.
